+++ lib/utils/platform_paths.dart
@@ -66,22 +66,3 @@ bool validateGameFolderPath(String newGameDir) {
     return false;
   }
 }
-
-bool validateIsProbablyAProgram(String path) {
-  try {
-    if (currentPlatform == TargetPlatform.macOS) {
-      if (path.toDirectory().existsSync()) {
-        return path
-            .toDirectory()
-            .resolve("Contents/MacOS")
-            .toDirectory()
-            .existsSync();
-      }
-    }
-
-    return path.toFile().existsSync();
-  } catch (e) {
-    Fimber.w("Error validating path as program", ex: e);
-    return false;
-  }
-}