+++ lib/widgets/under_construction_overlay.dart
@@ -2,6 +2,7 @@ import 'dart:math';
 
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
+import 'package:trios/widgets/moving_tooltip.dart';
 
 class UnderConstructionOverlay extends StatelessWidget {
   final Widget child;
@@ -10,15 +11,19 @@ class UnderConstructionOverlay extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Stack(
-      children: [
-        child,
-        Positioned.fill(
-          child: ClipRRect(
-            child: CustomPaint(painter: ConstructionTapePainter()),
+    return MovingTooltipWidget.text(
+      message: "UNDER CONSTRUCTION",
+      warningLevel: TooltipWarningLevel.warning,
+      child: Stack(
+        children: [
+          child,
+          Positioned.fill(
+            child: ClipRRect(
+              child: CustomPaint(painter: ConstructionTapePainter()),
+            ),
           ),
-        ),
-      ],
+        ],
+      ),
     );
   }
 }